Added search engine for events and actions

+++ src/main/java/com/kittyvt/domain/bnk/hirc/body/soundtype/events/Event.java
@@ -16,6 +16,6 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Event {
 
     @JsonProperty("actions")
-    public List<Integer> actions;
+    public List<Long> actions;
 
 }
